requirements.txt
discord
python-dotenv
asyncio
schedule
pytz
pandas
openpyxl
xlwt
xlrd